requirements.txt
transformers
sentencepiece
torch
tqdm
pytorch-lightning
wandb
nltk
scipy
scikit-learn
pandas
numpy
matplotlib
seaborn
rouge-score
nlp
torchmetrics
pytorch-lightning-bolts
pytorch-lightning
django
pytube3
whisper
jiwer
openai
google-api-python-client
python-dotenv